new specific config for api tests
